refactor(overlay): split main.rs along its four concerns
At 1044 lines the binary held the GPUI view, the agent conversation, the
click-away and role machinery, and startup wiring in one file — well past
the point where the repo splits into modules, and past the point where the
ring's async retry state machine can be read next to the window code.

- `ring` — the view and the window it opens, including cursor-centred
  placement clamped to the display the ring came up on.
- `agent` — connecting, receiving invocations, and the reporting pipeline
  with its supersede-aware retry.
- `session` — the click-away monitor and the `succession` role claim: the
  two things that decide when to stop showing a ring, or stop running.
- `main` — startup wiring only, at 130 lines.

The split exposed one piece of encapsulation worth keeping: dismissing a
ring reached into `RingView`'s private fields. Rather than widening them,
the view now answers `session_id()` and cancels itself, so the click-away
policy stays in `session` and the ring's state stays the ring's. `RingView`
and `Ipc` gained constructors for the same reason.

Tests moved to the module each exercises; all thirteen still pass, which is
what makes a move this mechanical safe to trust.
